use LONGLONG/ULONGLONG instead of 'long long' etc

+++ com/win32com/src/include/PythonCOM.h
@@ -704,7 +704,7 @@ public:
 		void *m_pValueHolder;
 		short m_sBuf;
 		long m_lBuf;
-		long long m_llBuf;
+		LONGLONG m_llBuf;
 		VARIANT_BOOL m_boolBuf;
 		double m_dBuf;
 		float m_fBuf;
